clean code: Rename `utils` to `hierarchy`. The Hierarchy support arranging data into a hierarchical data structure, throwing away entities when done loading and mapping one "level".
