+++ src/main/java/org/opentripplanner/netex/index/hierarchy/HierarchicalElement.java
@@ -1,4 +1,4 @@
-package org.opentripplanner.netex.index.util;
+package org.opentripplanner.netex.index.hierarchy;
 
 import javax.annotation.Nullable;
