+++ src/main/java/org/opentripplanner/netex/index/hierarchy/ReadOnlyHierarchicalVersionMapById.java
@@ -1,4 +1,4 @@
-package org.opentripplanner.netex.index.util;
+package org.opentripplanner.netex.index.hierarchy;
 
 import java.util.Collection;
