+++ src/main/java/org/opentripplanner/netex/mapping/ServiceLinkMapper.java
@@ -8,9 +8,9 @@ import org.opentripplanner.graph_builder.DataImportIssueStore;
 import org.opentripplanner.graph_builder.issues.MissingProjectionInServiceLink;
 import org.opentripplanner.model.FeedScopedId;
 import org.opentripplanner.model.ShapePoint;
-import org.opentripplanner.netex.index.util.ReadOnlyHierarchicalMap;
-import org.opentripplanner.netex.index.util.ReadOnlyHierarchicalMapById;
-import org.opentripplanner.netex.index.util.ReadOnlyHierarchicalVersionMapById;
+import org.opentripplanner.netex.index.hierarchy.ReadOnlyHierarchicalMap;
+import org.opentripplanner.netex.index.hierarchy.ReadOnlyHierarchicalMapById;
+import org.opentripplanner.netex.index.hierarchy.ReadOnlyHierarchicalVersionMapById;
 import org.rutebanken.netex.model.JourneyPattern;
 import org.rutebanken.netex.model.LinkInLinkSequence_VersionedChildStructure;
 import org.rutebanken.netex.model.LinkSequenceProjection_VersionStructure;
