+++ src/main/java/org/opentripplanner/netex/mapping/StopAndStationMapper.java
@@ -7,7 +7,7 @@ import org.opentripplanner.model.FareZone;
 import org.opentripplanner.model.Station;
 import org.opentripplanner.model.Stop;
 import org.opentripplanner.model.impl.EntityById;
-import org.opentripplanner.netex.index.util.ReadOnlyHierarchicalVersionMapById;
+import org.opentripplanner.netex.index.hierarchy.ReadOnlyHierarchicalVersionMapById;
 import org.opentripplanner.netex.support.StopPlaceVersionAndValidityComparator;
 import org.rutebanken.netex.model.Quay;
 import org.rutebanken.netex.model.Quays_RelStructure;
