+++ src/main/java/org/opentripplanner/netex/mapping/StopTimesMapper.java
@@ -6,7 +6,7 @@ import org.opentripplanner.model.StopLocation;
 import org.opentripplanner.model.StopTime;
 import org.opentripplanner.model.Trip;
 import org.opentripplanner.model.impl.EntityById;
-import org.opentripplanner.netex.index.util.ReadOnlyHierarchicalMap;
+import org.opentripplanner.netex.index.hierarchy.ReadOnlyHierarchicalMap;
 import org.opentripplanner.util.OTPFeature;
 import org.rutebanken.netex.model.DestinationDisplay;
 import org.rutebanken.netex.model.JourneyPattern;
