+++ src/test/java/org/opentripplanner/netex/index/hierarchy/E.java
@@ -1,4 +1,4 @@
-package org.opentripplanner.netex.index.util;
+package org.opentripplanner.netex.index.hierarchy;
 
 import org.rutebanken.netex.model.EntityInVersionStructure;
 import org.rutebanken.netex.model.EntityStructure;
