+++ src/test/java/org/opentripplanner/netex/index/hierarchy/HierarchicalElementTest.java
@@ -1,4 +1,4 @@
-package org.opentripplanner.netex.index.util;
+package org.opentripplanner.netex.index.hierarchy;
 
 import org.junit.Test;
