+++ src/test/java/org/opentripplanner/netex/index/hierarchy/HierarchicalMapTest.java
@@ -1,4 +1,4 @@
-package org.opentripplanner.netex.index.util;
+package org.opentripplanner.netex.index.hierarchy;
 
 import org.junit.Before;
 import org.junit.Test;
@@ -13,10 +13,10 @@ import static org.junit.Assert.assertEquals;
 import static org.junit.Assert.assertFalse;
 import static org.junit.Assert.assertNull;
 import static org.junit.Assert.assertTrue;
-import static org.opentripplanner.netex.index.util.E.EASTWOOD;
-import static org.opentripplanner.netex.index.util.E.REAGAN;
-import static org.opentripplanner.netex.index.util.E.SCHWARZENEGGER;
-import static org.opentripplanner.netex.index.util.SetSupport.sort;
+import static org.opentripplanner.netex.index.hierarchy.E.EASTWOOD;
+import static org.opentripplanner.netex.index.hierarchy.E.REAGAN;
+import static org.opentripplanner.netex.index.hierarchy.E.SCHWARZENEGGER;
+import static org.opentripplanner.netex.index.hierarchy.SetSupport.sort;
 
 /**
  * To test the hierarchical map we construct a hierarchy of:
