+++ src/test/java/org/opentripplanner/netex/mapping/MappingSupport.java
@@ -23,7 +23,7 @@ public class MappingSupport {
      * @see #createRef(String, Class)
      * @see #createJaxbElement(Object)
      */
-    static <T extends VersionOfObjectRefStructure> JAXBElement<T> createWrappedRef(String id, Class<T> clazz) {
+    public static <T extends VersionOfObjectRefStructure> JAXBElement<T> createWrappedRef(String id, Class<T> clazz) {
         return createJaxbElement(createRef(id, clazz));
     }
 
@@ -34,7 +34,7 @@ public class MappingSupport {
      * @param <T> the type of the created ref structure
      */
     @SuppressWarnings("unchecked")
-    static <T extends VersionOfObjectRefStructure> T createRef(String id, Class<T> clazz) {
+    public static <T extends VersionOfObjectRefStructure> T createRef(String id, Class<T> clazz) {
         try {
             return (T)clazz.newInstance().withRef(id);
         } catch (Exception e) {
@@ -48,7 +48,8 @@ public class MappingSupport {
      * @param <T> the value type
      * @return the value wrapped in a JAXBElement
      */
-    static <T> JAXBElement<T> createJaxbElement(@NotNull T value) {
+    @SuppressWarnings("unchecked")
+    public static <T> JAXBElement<T> createJaxbElement(@NotNull T value) {
         return new JAXBElement<T>(new QName("x"), (Class<T>)value.getClass(), value);
     }
 }
