+++ src/test/java/org/opentripplanner/netex/mapping/NoticeAssignmentMapperTest.java
@@ -8,8 +8,8 @@ import org.opentripplanner.model.StopTime;
 import org.opentripplanner.model.TransitEntity;
 import org.opentripplanner.model.Trip;
 import org.opentripplanner.model.impl.EntityById;
-import org.opentripplanner.netex.index.util.HierarchicalMapById;
-import org.opentripplanner.netex.index.util.HierarchicalMultimap;
+import org.opentripplanner.netex.index.hierarchy.HierarchicalMapById;
+import org.opentripplanner.netex.index.hierarchy.HierarchicalMultimap;
 import org.rutebanken.netex.model.MultilingualString;
 import org.rutebanken.netex.model.Notice;
 import org.rutebanken.netex.model.NoticeAssignment;
